test(security): move 13 test files into modules/security/__tests__/ (#707)
Relocate all security-domain tests alongside their module. Update
vi.mock paths and import references in both moved and consumer tests.
