+++ backend/src/modules/security/__tests__/ebpf-coverage.test.ts
@@ -13,14 +13,14 @@ const mockDb = {
 };
 
 // Kept: DB mock — tests assert SQL query patterns
-vi.mock('../core/db/app-db-router.js', () => ({
+vi.mock('../../../core/db/app-db-router.js', () => ({
   getDbForDomain: () => mockDb,
 }));
 
-import * as portainerClient from '../core/portainer/portainer-client.js';
-import * as portainerCache from '../core/portainer/portainer-cache.js';
-import { cache } from '../core/portainer/portainer-cache.js';
-import { closeTestRedis } from '../test-utils/test-redis-helper.js';
+import * as portainerClient from '../../../core/portainer/portainer-client.js';
+import * as portainerCache from '../../../core/portainer/portainer-cache.js';
+import { cache } from '../../../core/portainer/portainer-cache.js';
+import { closeTestRedis } from '../../../test-utils/test-redis-helper.js';
 import {
   getEndpointCoverage,
   updateCoverageStatus,
@@ -34,7 +34,7 @@ import {
   disableBeyla,
   enableBeyla,
   removeBeylaFromEndpoint,
-} from './ebpf-coverage.js';
+} from '../services/ebpf-coverage.js';
 
 let mockGetContainers: any;
 let mockGetEndpoints: any;
