+++ backend/src/modules/security/__tests__/harbor-vulnerability-store.test.ts
@@ -1,11 +1,11 @@
 import { beforeAll, afterAll, beforeEach, describe, it, expect, vi } from 'vitest';
-import { getTestDb, truncateTestTables, closeTestDb } from '../core/db/test-db-helper.js';
-import type { AppDb } from '../core/db/app-db.js';
+import { getTestDb, truncateTestTables, closeTestDb } from '../../../core/db/test-db-helper.js';
+import type { AppDb } from '../../../core/db/app-db.js';
 
 let testDb: AppDb;
 
 // Kept: app-db-router mock — redirects to test PostgreSQL instance
-vi.mock('../core/db/app-db-router.js', () => ({
+vi.mock('../../../core/db/app-db-router.js', () => ({
   getDbForDomain: () => testDb,
 }));
 
@@ -22,7 +22,7 @@ import {
   getLatestSyncStatus,
   cleanupOldVulnerabilities,
   type VulnerabilityInsert,
-} from './harbor-vulnerability-store.js';
+} from '../services/harbor-vulnerability-store.js';
 
 const makeVuln = (overrides: Partial<VulnerabilityInsert> = {}): VulnerabilityInsert => ({
   cve_id: 'CVE-2024-0001',
