+++ backend/src/modules/security/__tests__/pcap-analysis-service.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { parseTcpdumpOutput, buildAnalysisPrompt, parseAnalysisResponse } from './pcap-analysis-service.js';
-import type { PcapSummary } from '../core/models/pcap.js';
+import { parseTcpdumpOutput, buildAnalysisPrompt, parseAnalysisResponse } from '../services/pcap-analysis-service.js';
+import type { PcapSummary } from '../models/pcap.js';
 
 describe('parseTcpdumpOutput', () => {
   it('parses TCP packets with source/dest IP and ports', () => {
