+++ backend/src/modules/security/__tests__/security-audit-route.test.ts
@@ -1,9 +1,9 @@
 import { afterAll, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 import Fastify, { type FastifyInstance, type FastifyReply, type FastifyRequest } from 'fastify';
 import { validatorCompiler, serializerCompiler } from 'fastify-type-provider-zod';
-import { getTestDb, truncateTestTables, closeTestDb } from '../core/db/test-db-helper.js';
-import type { AppDb } from '../core/db/app-db.js';
-import { monitoringRoutes } from './monitoring.js';
+import { getTestDb, truncateTestTables, closeTestDb } from '../../../core/db/test-db-helper.js';
+import type { AppDb } from '../../../core/db/app-db.js';
+import { monitoringRoutes } from '../../../routes/monitoring.js';
 
 let testDb: AppDb;
 
@@ -21,7 +21,7 @@ vi.mock('../services/security-audit.js', () => ({
 }));
 
 // Kept: app-db-router mock — tests control database routing
-vi.mock('../core/db/app-db-router.js', () => ({
+vi.mock('../../../core/db/app-db-router.js', () => ({
   getDbForDomain: () => testDb,
 }));
